IIS_ITU/proj1/src/web/backend/requirements.txt
Flask==2.2.2
Flask_APScheduler==1.12.1
Flask_Cors==3.0.10
flask_marshmallow==0.14.0
flask_login==0.6.2
Flask_SQLAlchemy==2.5.1
marshmallow==3.19.0
requests==2.28.1
SQLAlchemy==1.4.44
marshmallow-sqlalchemy==0.28.1
psycopg2-binary==2.9.4
werkzeug==2.2.2
pyjwt
gunicorn